requirements.txt
python-dotenv>=1.0.0
requests>=2.28.0
sv_ttk>=0.3.0
Pillow>=9.0.0
googletrans==4.0.0-rc1